Separated charting sink; added rsocket server for chart data; more separation of work in charting/strategy classes; finally broke down and used reflection to see inside CCI data.

+++ src/main/java/com/jessethouin/strategy/charts/AbstractChart.java
@@ -52,7 +52,7 @@ public abstract class AbstractChart extends ApplicationFrame {
         this.pack();
         UIUtils.centerFrameOnScreen(this);
         this.setVisible(true);
-        alpacaChartDataSink.subscribe(chartData -> addChartData(chartData.getClose().floatValue()));
+        alpacaChartDataSink.subscribe(chartData -> addChartData(chartData.getClose()));
     }
 
     protected void addChartData(float close) {
